Putting remove_structure in contents.remove_structure submodule.
Also adding a unit test for this and prettyprint.
